expose a single function for the component
